% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.3g.R
\name{cplhood}
\alias{cplhood}
\title{Conditional pseudo-likelihood for a set of observations of Z_d and Z_a. Objective function for fit.cond}
\usage{
cplhood(Z, pars, weights = rep(1, dim(Z)[1]), C = 1)
}
\arguments{
\item{Z}{an n x 2 array; Z[i,1], Z[i,2] are the Z_d and Z_a scores respectively for the ith SNP}

\item{pars}{vector containing initial values of pi0,pi1,tau,sigma1,sigma2,rho.}

\item{weights}{SNP weights to adjust for LD; output from LDAK procedure}

\item{C}{scaling factor for adjustment to ensure model identifiability}
}
\value{
value of pseudo- log likelihood of all observations (if sumlog==TRUE) or vector of n pseudo-likelihoods (if sumlog==FALSE)
}
\description{
Conditional pseudo-likelihood for a set of observations of Z_d and Z_a. Objective function for fit.cond
}
\author{
James Liley
}
