% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{d_par}
\alias{d_par}
\title{d_par}
\usage{
d_par(pars1, ...)
}
\arguments{
\item{pars1}{an n x 6 matrix of parameters}

\item{...}{additional parameters passed to dist}
}
\value{
an n x m matrix A such that A[i,j] is the distance between pars1[i,] and pars2[j,]
}
\description{
Distance metric between sets of parameters. Computed as a Euclidean distance in a transformed parameter space. pi0-pi2 are transformed to their inverse logits, as a difference 0.001-0.1 means more than a difference 0.4-0.5. Tau, sigma1, and sigma2 are transformed to sqrt(|x-1|), as differences are most important for values near 1. Rho is transformed to x^2 /2, since differences are only important for large rho.
}
\examples{
px=pars_start(N=5)
dx=d_par(px)
image(as.matrix(dx))
}
\author{
James Liley
}

