% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{dataset_values}
\alias{dataset_values}
\title{Estimate the values of ... and ... suggested by a set of \code{\link{Zd}} and \code{\link{Za}} values.}
\usage{
dataset_values(Za, Zd)
}
\description{
Estimate the values of ... and ... suggested by a set of \code{\link{Zd}} and \code{\link{Za}} values.
}

