% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.3g.R
\name{fit.2g}
\alias{fit.2g}
\title{fit.2g}
\usage{
fit.2g(Z, pars = c(0.5, 1), weights = rep(1, length(Z)), C = 1, sgm = 0,
  ...)
}
\arguments{
\item{Z}{a vector of length n}

\item{pars}{vector containing initial values of \code{pi0}, \code{s}.}

\item{weights}{SNP weights to adjust for LD; output from LDAK procedure}

\item{C}{a term \eqn{C log(}\code{pi0}*\code{1-pi0}\eqn{)} is added to the likelihood so the model is specified.}

\item{sgm}{force \code{s} \eqn{\ge sgm}}

\item{...}{other parameters passed to R's \code{\link{optim}} function}
}
\value{
a list of three objects: \code{pars} is the vector of fitted parameters under H1, \code{h1value} is the pseudo-likelihood under H1, \code{h0value} is the pseudolikelihood under H0.
}
\description{
Fit a specific one-dimensional Gaussian mixture distribution to weighted SNP data
}
\details{
\deqn{
  pdf(Z) = \pi_{0} N(0,1) + (1-\pi_{0}) N(0,s^{2})
 }{
  Z_a,Z_d ~ pi0 N(0,1) + (1-pi0) N(0,s^2)
 }

The model is characterised by the vector \code{\link{pars}}=(\code{pi0},\code{s}). Under the null hypothesis that all SNPs are null, pi0=0.5, s=1.

This function finds the maximum pseudo-likelihood estimators for the paramaters of these three Gaussians, and the mixing parameters representing the proportion of SNPs in each category.
}
\examples{
nn=100000
Z=c(rnorm(0.8*nn), rnorm(0.2*nn,sd=3)); weights=runif(nn)
yy=fit.2g(Z,pars=c(0.5,1),weights=weights)
yy$pars
}
\author{
Chris Wallace and James Liley
}

