% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{fit.em1}
\alias{fit.em1}
\title{fit.em1}
\usage{
fit.em1(Z, weights = rep(1, length(Z)), pi0 = 0.95, sigma_init = 2,
  tol = 1e-04, verbose = TRUE, maxit = 10000)
}
\arguments{
\item{Z}{numeric vector of observed data}

\item{sigma_init}{initial value for \code{\link{sigma}}}

\item{tol}{how small a change lhood prompts continued optimization}

\item{maxit}{maximum number of iterations}

\item{pi0_init}{value for \code{\link{pi0}}}
}
\value{
a list containing fitted \code{\link{pi0}}, fitted \code{\link{sigma}}, and a record of fitted parameters at eachs stage of the E-M algorithm.
}
\description{
Fit a specific two Guassian mixture distribution to a set of Z values, as for \code{\link{fit.em}} but fixing \code{pi0} at the value of \code{\link{pi0_init}}, only fitting \code{sigma}.
}
\details{
Assumes 
Z ~ N(0,1) with probability  \code{\link{pi0}}, Z ~ N(0,1 + \code{\link{sigma}}^2) with probability 1-\code{\link{pi0}}
}
\examples{
sigma <- 2
pi0 <- 0.8
n <- 10000; n0=round(pi0*n); n1=n-n0
Z <- c(rnorm(n0,0,1),rnorm(n1,0,sqrt(1+ (sigma^2))))
fit<-fit.em(Z)
fit$pi0
fit$sigma
fit$history
}
\author{
Chris Wallace, James Liley
}

