% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_snps.R
\name{getpx}
\alias{getpx}
\title{getpx}
\usage{
getpx(Z, pars)
}
\arguments{
\item{Z}{n x 2 matrix of Z scores; Z[,1]=Z_d, Z[,2]=Z_a}

\item{pars}{parameters of full model; in order, pi0, pi1, tau, sigma_1, sigma_2, rho}

\item{i}{get posterior probability of this group; if null, get posterior probability of all groups.}
}
\value{
n x 3 matrix; element [i,j] is probability of SNP i being in group j.
}
\description{
Given some set of parameters in pars, compute the posterior likelihood of membership of each group for a set of SNP Z scores
}
\examples{
nn=100000
Z=abs(rbind(rmnorm(0.8*nn,varcov=diag(2)), rmnorm(0.15*nn,varcov=rbind(c(1,0),c(0,2^2))), rmnorm(0.05*nn,varcov=rbind(c(3^2,2),c(2,4^2)))));
pars=c(0.8,0.15,3,2,4,2)
gp=getpx(Z,pars)
plotZ(Z,rlim=2,col=rgb(gp))
}
\author{
James Liley
}

