% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.3g.R
\name{grad_logl}
\alias{grad_logl}
\title{grad_logl}
\usage{
grad_logl(Z, pars, weights = rep(0, dim(Z)[1]), C = 1)
}
\arguments{
\item{Z}{an n x 2 array; Z[i,1], Z[i,2] are the \eqn{Z_{d}}{Z_d} and \eqn{Z_{a}}{Z_a} scores respectively for the \eqn{i}th SNP}

\item{pars}{vector containing initial values of (\code{pi0},\code{pi1},\code{tau},\code{sigma1},\code{sigma2},\code{rho}).}

\item{weights}{SNP weights to adjust for LD; output from LDAK procedure}

\item{C}{a term \eqn{C*log(}\code{pi0}*\code{pi1}*\code{pi2}\eqn{)} is added to the likelihood so the model is specified.}
}
\value{
Vector of partial derivatives of pseudo-log-likelihood with respect to each of the parameters in \code{\link{pars}}
}
\description{
Compute gradient of pseudo-log likelihood for a dataset at given parameters
}
\examples{
nn=100000
Z=abs(rbind(rmnorm(0.8*nn,varcov=diag(2)), rmnorm(0.15*nn,varcov=rbind(c(1,0),c(0,2^2))), rmnorm(0.05*nn,varcov=rbind(c(3^2,2),c(2,4^2))))); weights=runif(nn)
grad(Z,pars=c(0.7,0.2,2.5,1.5,3,1),weights=weights,C=1)
yy$pars
plot(yy,rlim=2)
}
\author{
James Liley
}
