% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{intet}
\alias{intet}
\title{intet}
\usage{
intet(x, p1, p2, p3, p4)
}
\arguments{
\item{x}{point to be tested (three element vector)}

\item{p1}{first vertex of tetrahedron (three element vector)}

\item{p2}{second vertex of tetrahedron (three element vector)}

\item{p3}{third vertex of tetrahedron (three element vector)}

\item{p4}{fourth vertex of tetrahedron (three element vector)}
}
\value{
TRUE/FALSE
}
\description{
Tests if a point is inside a tetredron
}
\examples{
 p1=c(0,0,2); p2=c(0,2,0); p3=c(2,0,0); p4=c(3,3,3); x=c(2,2,2); y=c(1,1,0)
 intet(x,p1,p2,p3,p4)
 intet(y,p1,p2,p3,p4)
}
