% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{maketab}
\alias{maketab}
\title{maketab}
\usage{
maketab(sims, p)
}
\arguments{
\item{sims}{matrix of simulation results. See above for necessary form.}

\item{p}{tabulate power to detect this level of significance}
}
\value{
matrix with dimensions n x 4; each row takes the form: n3, tau, sigma2, power.
}
\description{
Given a set of simulations for power calculations, tabulate power as a function of values of the number of SNPs in category 3 (n3), sigma2, and tau.
}
\details{
Each row of the matrix of simulations must have length >25, in which the first 18 elements are identical to those from general simulation output (ie: likelihood under H1, likelihood under H0, correcting factor, number of iterations to fit H1 model, number of iterations to fit H0 model, random seed, parameters of H0 model, parameters of H1 model) followed by: number of SNPs in category 1, number of SNPs in category 2, number of SNPs in category 3, true value of tau, true value of sigma1, true value of sigma2, true value of rho. Other matrix columns are ignored.

Simulations must be such that the asymptotic distribution of PLR under the null is a mixture of chi^2 distributions with one and two degrees of freedom, where distributions have equal weights.
}

