% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulations.R
\name{mixchipars}
\alias{mixchipars}
\title{mixchipars}
\usage{
mixchipars(plr, init = c(1, 0.5))
}
\arguments{
\item{plr}{set of plr values}

\item{init}{values of gamma, kappa at which to start numerical fitting algorithm}
}
\value{
two element vector; gamma, kappa.
}
\description{
Fit maximum-likelihood values of gamma and kappa to a set of PLR values
}
\details{
Assumes that PLR has a mixture distribution with some weight p0 at 0, a weight 'kappa' for a distribution gamma*chi_1^2, and a weight 1-kappa-p0 for a distribution gamma*chi_2^2.
}
\examples{
gamma=0.7;kappa=0.4;p0=0.05; N=10000
px=sort(gamma*c(rep(0,p0*N),rchisq(kappa*N,df=1),rchisq((1-kappa)*N,df=2)))
mixchipars(px)
}

