% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{npower}
\alias{npower}
\title{npower}
\usage{
npower(sig_u = NULL, sig_o = NULL, n1 = NULL, n2 = NULL)
}
\arguments{
\item{sig_u}{underlying standard deviation of odds ratio distribution. Estimable using function \code{\link{dataset_values}}. Set to NULL to compute from other values.}

\item{n1}{number of cases in subgroup/group 1. Set to NULL to compute conditional on other values. If both \code{\link{n1}} and \code{\link{n2}} are NULL, computes required number of cases conditional on \code{\link{n1}}=\code{\link{n2}}}

\item{n2}{number of cases in subgroup/group 2. Set to NULL to compute conditional on other values. If both \code{\link{n1}} and \code{\link{n2}} are NULL, computes required number of cases conditional on \code{\link{n1}}=\code{\link{n2}}}
}
\description{
Compute number of cases in order that observed value of sigma (that is, \code{tau}, \code{s2}) reaches a given value.
}
\details{
The observed standard deviation depends on the underlying distribution of population odds ratios; that is, odds ratios between population minor allele frequencies of the causative and protective alleles. This can be estimated using the function \code{\link{dataset_values}}

This function computes the number of cases in two groups necessary to reach a given observed standard deviation (\code{\link{sig_o}}) for effect sizes between then (\code{tau} or \code{s2}). The two groups correspond to either the number of cases and the number of controls (in which case \code{\link{sig_o}}=\code{s2}) or the number of cases in each subgroup (in which case \code{\link{sig_o}}=\code{tau}).
}
