% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{Nsim_stat}
\alias{Nsim_stat}
\title{Nsim_stat}
\usage{
Nsim_stat(lr, NN = c(50, 100, 500, 1000, 1500), NB = 500, p_at = c(2, 3,
  5, 10), quant_at = pnorm(-2:2), verbose = TRUE)
}
\arguments{
\item{lr}{matrix of simulation results}

\item{NN}{vector of integer; test performance of simulation fit for n simulations, where n takes the values in NN}

\item{NB}{number of bootstrap samples to take for each sample size in NN}

\item{p_at}{compute p values at each value in p_at}

\item{quant_at}{for each size in NN and each parameter, record these quantiles of the bootstrapped parameter}

\item{verbose}{print values of NN[i] as they complete.}
}
\value{
list of two objects, stat and quant; stat[,ii,i] is, in order, kappa, gamma, and p-values given gamma and kappa of each value in p_at for the i'th bootstrap sample for bootstrap sample size NN[ii]; quant[k,ii,j] is the quant_at[j]'th quantile of boostrapped test statistic k amongst bootstrap samples of size NN[ii]
}
\description{
Computes how well the parameters gamma and kappa are fitted at various numbers of simulations, and how these affect the computation of p-values. Returns an object of class 'sim_N_test' which can be plotted.
}
\examples{
data("sim_mat.RData") 
stat=Nsim_stat(sim_mat)
summary(stat)
plot(stat)
}
\author{
James Liley
}
