% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{p_value}
\alias{p_value}
\title{p_value}
\usage{
p_value(x, cplr)
}
\arguments{
\item{x}{list of plr values at which to compute p value}

\item{cplr}{list of cplr values from simulations from which to estimate null distribution}
}
\description{
Computes the p-value of an observed PLR according to either empirical quantile in simulations, a mixture chi-square distribution, or the empirical bound, depending on size.
}
\author{
James Liley
}
