% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/processing.R
\name{pars_start}
\alias{pars_start}
\title{pars_start}
\usage{
pars_start(Z, weights = rep(1, dim(Z)[1]), H = 1, n1 = 1000, n2 = 10,
  n3 = 3, C = 1, nx = 10, seed = NULL)
}
\arguments{
\item{Z}{an n x 2 array; Z[i,1], Z[i,2] are the Z_d and Z_a scores respectively for the ith SNP}

\item{weights}{SNP weights to adjust for LD; output from LDAK procedure}

\item{H}{hypothesis, 0 or 1}

\item{n1}{begin with n1 random parameter sets from pars_rand, chosen according to prior distributions of parameters (see documentation for pars_rand)}

\item{n2}{trim initial list to this many well-separated sets of parameters; each of these parameter sets undergoes nx steps of the EM algorithm}

\item{n3}{finally take n3 well-separated sets of parameters for entry into the final E-M algorithm}

\item{C}{scaling factor for adjustment}

\item{nx}{use this many iterations of the E-M algorithm on each of the n2 sets of paramaters}

\item{seed}{random seed for generating results. Use to regenerate.}
}
\description{
Obtain a set of starting parameters for E-M algorithm, given Z_a and Z_d scores and LDAK weights. Because of the potential for local maxima in the likeilhood landscape, it is important to start the algorithm at several points. Since the E-M algorithm is computationally intensive, it is useful to start as close as possible to the actual maxima, and at as few points as possible. This function seeks a small number of 'promising' start points with high likelihood, sufficiently far from each other.
}
\author{
James Liley
}

