% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulations.R
\name{pbind}
\alias{pbind}
\title{pbind}
\usage{
pbind(x, S = NULL, plr = NULL, tau = NULL, s1 = NULL, s2 = NULL,
  rho = NULL, res = 1000)
}
\arguments{
\item{x}{list of plr values at which to compute p value}

\item{S}{object of class sim_out, containing PLR, tau, s1, s2, rho. Overrides other parameters if set.}

\item{plr}{list of plr values from simulation}

\item{tau}{list of fitted values of tau corresponding to values of plr}

\item{s1}{list of fitted values of sigma_1 corresponding to values of plr}

\item{s2}{list of fitted values of sigma_2 corresponding to values of plr}

\item{rho}{list of fitted values of rho corresponding to values of plr}

\item{res}{resolution of numerical integral}
}
\description{
Computes a bound on the p-value for a subtype comparison, relatively robust to LD, by principally using fitted values rather than overall likelihood.
}
\details{
The bound is very conservative for most observed pseudo-likelihood ratios. However, at very high likelihood ratios, the mixture-chi squared approximation appears to break down, meaning that p-values generated using this algorithm could result in underestimated false-discovery rates. We recommend that this function is used to estimate p-values if the observed PLR is more than twice the maximum obtained in simulated data, or for PLR greater than the maximum simulated value if the observed distribution of PLR statistics appears to poorly match the mixture-chi squared distribtion at high PLR values.
}
\examples{
lr=read.table("~/Subtypes/AI_data/simulation/LR");
plr=lr[,2]-lr[,1]-lr[,3]; plr[which(plr<0)]=0; tau=lr[,34]; s1=lr[,36]; s2=lr[,37]; rho=lr[,38]
pbind(c(0.5,1,1.5,2,3,5,10,20,30),plr,tau,s1,s2,rho)
}
\author{
James Liley
}

