% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit.mml5.R
\name{plhood1}
\alias{plhood1}
\title{plhood1}
\usage{
plhood1(za, pars, weights = rep(1, length(za)), C = 1, sumlog = TRUE)
}
\arguments{
\item{za}{one dimensional vector of Z_a scores}

\item{pars}{vector of parameters (pi0,pi1,tau,sigma1,sigma2,rho)}

\item{weights}{LDAK weights corresponding to Z_a}

\item{C}{scaling factor for adjustment}

\item{sumlog}{set to TRUE to return the (weighted) sum of log pseudo-likelihoods for each datapoint; FALSE to return a vector of pseudo-likelihoods for each datapoint.}
}
\value{
value of pseudo- log likelihood of all observations (if sumlog==TRUE) or vector of n pseudo-likelihoods (if sumlog==FALSE)
}
\description{
Computes the likelihood of Z_a alone; used for establishing how much of a likelihood ratio is because Z_a is better-fitted by a three-Gaussian model. Strictly, calculates the expected value of Z_a/Z_d when Z_a has uniform unit variance, and the parameter values tau, rho are set to 1, 0 respectively.
}
\author{
James Liley
}

