% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit.mml5.R
\name{plot.3Gfit}
\alias{plot.3Gfit}
\title{Plot method for class 3Gfit. If the z_ad attribute of yy is set, then plots the absolute Z_a and Z_d scores (+/+ quadrant only), blocking out dense regions, and contours of the fitted Gaussians. Note that group 3 comprises two mirror-image Gaussians. Contours for group 1 are drawn in black, group 2 in blue, and group 3 in red.}
\usage{
plot.3Gfit(yy, ...)
}
\arguments{
\item{yy}{object of class 3Gfit; generally output from fit.3g}

\item{scales}{set to draw contours of the distributions at these levels (passed to plotpars and plotvdist).}
}
\description{
Plot method for class 3Gfit. If the z_ad attribute of yy is set, then plots the absolute Z_a and Z_d scores (+/+ quadrant only), blocking out dense regions, and contours of the fitted Gaussians. Note that group 3 comprises two mirror-image Gaussians. Contours for group 1 are drawn in black, group 2 in blue, and group 3 in red.
}
\author{
James Liley
}

