% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.sim_N_test}
\alias{plot.sim_N_test}
\title{summary.Nsim_stat}
\usage{
plot.sim_N_test(yy, all = FALSE)
}
\arguments{
\item{yy}{object of class sim_N_test, output from Nsim_stat}

\item{all}{combine all plots in one}
}
\description{
Plot method for class sim_N_test
}
\details{
Draws plot of values and quantiles of each test statistic for each bootstrap sample size
}
\author{
James Liley
}

