% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{plot.sim_output}
\alias{plot.sim_output}
\title{Plot method for class sim_output. Plots the values of PLR in ascending order along with mixture chi-squared distribution, and a Q-Q plot comparing the two. Sets global variables X,Y where (X,Y) are the points on the Q-Q plot.}
\usage{
plot.sim_output(yy, QQ = NULL, conf = FALSE, q95 = FALSE, ...)
}
\arguments{
\item{yy}{object of class sim_output; generally output from sim_analysis}

\item{QQ}{set to TRUE to only draw QQ plot, FALSE to only draw chi-squared plot, or NULL to draw both.}

\item{conf}{set to TRUE to draw a 95\% confidence interval on the Q-Q plot. Confidence intervals are constructed by repeatedly randomly sampling from a mixture chi-squared distribution parametrised by yy$gamma and yy$kappa. For the random sample, the best-fit values of gamma and kappa (gamma' and kappa') are recovered, and the random sample is compared to the fitted mixture-chi square parametrised by gamma' and kappa'. Finally, standard errors are estimated empirically at each quantile. Computing limits is slow, so if this option is TRUE, global variables lb and ub are set, where (X, lb), (X,ub) are the co-ordinates of the lower and upper confidence limits.}

\item{q95}{set to TRUE to draw lines on the plot corresponding to empirical type 1 error if generating a significance cutoff from mixture-chi squared corresponding to p<0.05.}

\item{...}{additional parameters passed to plot}
}
\description{
Plot method for class sim_output. Plots the values of PLR in ascending order along with mixture chi-squared distribution, and a Q-Q plot comparing the two. Sets global variables X,Y where (X,Y) are the points on the Q-Q plot.
}
\author{
James Liley
}

