% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit.mml5.R
\name{plotpars}
\alias{plotpars}
\title{plotpars}
\usage{
plotpars(pars, over = TRUE, ...)
}
\arguments{
\item{pars}{six element vector (pi0,pi1,tau,sigma1,sigma2,rho)}
}
\description{
Plots contour lines of three bivariate normal distributions as determined by 'pars'
}

