% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.3g.R
\name{plotvdist}
\alias{plotvdist}
\title{plotvdist}
\usage{
plotvdist(varcov = diag(2), m = c(0, 0), scales = (2^(-(1:5))),
  relative = TRUE, over = TRUE, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{varcov}{covariance matrix (2 x 2)}

\item{m}{mean vector}

\item{scales}{heights of normal PDF at which to draw contour lines. If \code{\link{relative}}=TRUE, draws contours at heights given by \code{\link{scales}} * 1 /(2 \code{pi} \code{|\link{varcov}|}); if \code{\link{relative}}=FALSE, draws contours at heights given by \code{\link{scales}} directly.}

\item{relative}{set to TRUE to draw heights relative to determinant of \code{\link{varcove}}. See parameter \code{\link{scales}}.}

\item{over}{set to FALSE to draw new plot; TRUE to overplot existing plot}

\item{xlim}{limits of x axis if \code{\link{over}}=FALSE}

\item{ylim}{limits of y axis if \code{\link{over}}=FALSE}
}
\description{
Plots contour lines of a bivariate normal distribution parametrised by covariance matrix, at specific distribution heights \code{\link{scales}}
}
\examples{
mat1=rbind(c(3,-1),c(-1,2));
mat2=rbind(c(1,0.8),c(0.8,1))
plotvdist(mat1,c(1,1),xlim=c(-5,5),ylim=c(-5,5),over=FALSE,col="red")
plotvdist(mat2,c(0,0),over=TRUE)
abline(h=0,lty=2); abline(v=0,lty=2)
}
