% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.3g.R
\name{plotZ}
\alias{plotZ}
\title{plotZ}
\usage{
plotZ(z_ad, rlim = quantile(z_ad[, 2], 0.99), col = rep("black",
  dim(z_ad)[1]), mcol = NULL, abs = (min(z_ad) < 0), over = FALSE, ...)
}
\arguments{
\item{rlim}{block out circle of radius \code{\link{rlim}}; only plot points with \eqn{Z_{d}^{2} + Z_{a}^{2} > \texttt{rlim}^{2}}{Z_d^2 + Z_a^2 > rlim^2}.}

\item{col}{colour of points or vector describing colour of each point of \code{\link{Z}}}

\item{mcol}{colour of middle blocked-out region; defaults to first element of \code{\link{col}}}

\item{abs}{set to TRUE to plot only upper right quadrant}

\item{over}{set to TRUE to plot over existing plot; FALSE to draw new plot.}

\item{Z}{an n x 2 array; Z[i,1], Z[i,2] are the \eqn{Z_{d}}{Z_d} and \eqn{Z_{a}}{Z_a} scores respectively for the \eqn{i}th SNP}
}
\description{
Plot \eqn{Z_{a}}{Z_a} and \eqn{Z_{d}}{Z_d} scores efficiently, blocking out dense region near origin.
}
\examples{
nn=100000
Z=abs(rbind(rmnorm(0.8*nn,varcov=diag(2)), rmnorm(0.15*nn,varcov=rbind(c(1,0),c(0,2^2))), rmnorm(0.05*nn,varcov=rbind(c(3^2,2),c(2,4^2)))));
plotZ(Z,rlim=2,col="red",cex=0.5)
}
\author{
James Liley
}

