% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\title{General power calculation method for PLR.}
\usage{
power(power = NULL, p = NULL, n3 = NULL, s2 = NULL, tau = NULL)
}
\arguments{
\item{p}{significance level at which to estimate power}

\item{s2}{underlying parameter for which power is to be estimated. Marginal standard deviation for \eqn{Z_a}{Z_{a}} in category 3}

\item{tau}{underlying parameter for which power is to be estimated. Marginal standard deviation for \eqn{Z_d}{Z_{d}} in category 3}

\item{ll}{object of class \code{\link{lm}}. Given simulated values, \code{\link{ll}} should be generated by a function of the form \code{ll=lm(I(log(1+plr)~ pi2+ poly(s2,2)+poly(tau,2) + I(pi2*s2) + I(pi2*tau))))}. Simulated values should be from uncorrelated observations on unweighted SNPs, so the asymptotic distribution of PLR is mixture-chi^2.

Power also depends on parameters \code{s1}, the marginal standard deviation for \eqn{Z_{a}}{Z_a} in category 2, \code{pi1}, the proportion of SNPs in category 2, and \code{rho}, the correlation between \eqn{Z_{a}}{Z_a} and \eqn{Z_{d}}{Z_d} in category 3, but the influence on power from these variables is small.}

\item{pi2}{underlying parameter for which power is to be estimated. Corresponds to number of SNPs in category 3}

\item{rho}{underlying parameter for which power is to be estimated. Correlation between \eqn{Z_a}{Z_{a}} and \eqn{Z_d}{Z_{d}} in category 3}
}
\description{
Power of the PLR method depends on the number of SNPs in category 3 (\code{\link{n3}}), and the marginal variances in category 3 of \eqn{Z_{a}}{Z_a} (\code{\link{s2}}) and \eqn{Z_{a}}{Z_a} (\code{\link{tau}}). Given a required power and p-value,
}
\details{
Function to estimate power to reject the null hypothesis at significance \code{\link{p}} given underlying parameters \code{\link{pi2}}, \code{\link{s2}}, \code{\link{tau}} and \code{\link{rho}}.
}
