% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_interpolate}
\alias{power_interpolate}
\title{power_interpolate}
\usage{
power_interpolate(x, ptab)
}
\arguments{
\item{x}{vector of values n3 (number of SNPs in category 3),tau,sigma2,rho at which to estimate power}

\item{ptab}{matrix of simulations; each row n3, tau, sigma2, power.}
}
\value{
estimate of power at new set of parameter values
}
\description{
Given a table of observed power values from simulations at a range of parameter sets, estimates power at a new set of parameters by linear interpolation. Table of power values must be complete.
}
