% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit.mml5.R
\name{print.3Gfit}
\alias{print.3Gfit}
\title{Print method for class 3Gfit}
\usage{
print.3Gfit(y, n = 3, ...)
}
\arguments{
\item{yy}{object of class 3Gfit; generally output from fit.3g}
}
\description{
Print method for class 3Gfit
}
\author{
James Liley
}

