% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulations.R
\name{print.sim_output}
\alias{print.sim_output}
\title{Print method for class sim_output}
\usage{
print.sim_output(yy)
}
\arguments{
\item{yy}{object of class sim_output; generally output from sim_analysis}
}
\description{
Print method for class sim_output
}
\author{
James Liley
}

