% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{rand_analysis}
\alias{rand_analysis}
\title{rand_analysis}
\usage{
rand_analysis(sims, maxit = 10000)
}
\arguments{
\item{sims}{either a matrix of outputs from sim_gen, a directory and file basename, or a list of files.}

\item{maxit}{if the E-M algorithm uses the maximum allowed number of iterations, it may not have converged. Set this parameter to restrict to only simulations taking fewer than this number of iterations.}
}
\description{
Given outputs of fitted parameters and likelihoods from random subtypes, analyse outcomes in order to compute p-values by extrapolation or interpolation.
}
\examples{
data(sim_mat)
S=sim_analysis(sim_mat)
plot(S)
}
\author{
James Liley
}

