% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulations.R
\name{rand_gen}
\alias{rand_gen}
\title{rand_gen}
\usage{
rand_gen(X, Z_a, pars_init1 = c(0.8, 0.15, 2, 1, 2, 1), pars_init0 = c(0.8,
  0.15, 2, 2, 1, 0), weights = rep(1, length(Z_a)), Ca = NULL, Cd = NULL,
  n0 = NULL, Yd = NULL, file = NULL, seed = NULL, ...)
}
\arguments{
\item{X}{a snpMatrix object, QC'd, or a list of objects of type snpMatrix covering different SNPs (ie, one per chromosome). X is assumed to only contain cases as rows, as it somewhat reduces efficiency to split the matrix every time.}

\item{Z_a}{Z_a scores (case vs control); these will be the same for each random subgroup (as the case set is the same) so it is inefficient to recalculate each time.}

\item{pars_init1}{an nx6 array of starting parameter values for fitting full model. The E-M algorithm is run from each of them in turn.}

\item{pars_init0}{an nx6 array of starting parameter values for fitting null model. The E-M algorithm is run from each of them in turn.}

\item{weights}{SNP weights to adjust for LD; output from LDAK procedure}

\item{Ca}{matrix of covariates for computing Z_a; Ca[i,] gives the covariates for sample i corresponding to X[i,]}

\item{Cd}{matrix of covariates for computing Z_d; Cd[i,] gives the covariates for sample i corresponding to X[i,]. CD[i,] is ignored unless Ya[i]>0.}

\item{n0}{optionally, only simulate subtypes such that the smaller subtype has size n0. If n0 is null, use random subgroup sizes with the smaller subtype has between 10% and 50% of total cases.}

\item{Yd}{optionally choose random subtyping by randomly sorting 'true' Yd. This is the best option for quantitative traits Yd. Supercedes n0 if set.}

\item{file}{set to a directory where output of the simulation will be saved. The output file is small but the simulation is likely to be repeated multiple times. If file is null, print output of the simulation to console.}

\item{seed}{random seed for computing random subtypes. Can be used to reconstruct simulation. By default, chosen by time.}

\item{...}{other arguments passed to fit.3g}
}
\description{
For a genotype matrix, simulate a random subtyping, compute corresponding Za and Zd scores, and fit model.
}
\examples{
# pending
}
\author{
James Liley
}

