% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genotype_sim.R
\name{s2p}
\alias{s2p}
\title{s2p}
\usage{
s2p(s, n1, n2, mafs = NULL, NS = 50000)
}
\arguments{
\item{s}{standard deviation of underlying log odds ratio distribution}

\item{n1}{number of samples in group 1}

\item{n2}{number of samples in group 2}

\item{mafs}{use to specify distribution of average MAF (across both groups). Default is MAF~U(0.01,0.5)}

\item{NS}{number of SNPs to use in simulation}
}
\value{
expected value of observed standard deviation of Z scores.
}
\description{
Inverse function for p2s. Converts expected values of SD(Z) (ie, tau, sigma_1, or sigma_2) into underlying 'population' odds-ratio distribution. Assume that, across some set of SNPs in, population log-odds ratios between two phenotypes A and B are normally distributed with standard deviation p. If a GWAS is performed between a group of samples with phenotype A of size n1 and a group of samples of phenotype B of size n2, and Z-scores calculated for each SNP with standard deviation s, this function recovers p.
}

