% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genotype_sim.R
\name{sim_genotypes}
\alias{sim_genotypes}
\title{sim_genotypes}
\usage{
sim_genotypes(n_snps = 50000, n_control = 2000, n_case = c(1000, 1000),
  pars = NULL, q2SEd = NULL, q2SEa1 = NULL, q2SEa2 = NULL,
  cor_st = NULL, seed = NULL, return_matrix = FALSE, null_model = TRUE)
}
\arguments{
\item{n_snps}{number of SNPs}

\item{n_control}{number of controls}

\item{n_case}{two-element vector; number of cases in subtypes 1 and 2}

\item{pars}{expected observed parameter values. If NULL, values are chosen randomly; pi0 from c(0.1,0.01,0.001),1), pi1 from c(0.2,0.1,0.05); if parameter q2SEd is null, q2SEd from c(1,1.2,1.5,2); if parameter q2SEa is null, q2SEa1 from c(1.2,1.5,2); if parameter q2SEa is null and null_model==FALSE, q2SEa2 from c(1.2,1.5,2); if parameter rho is null and null_model==TRUE, rho from c(0, 0)}

\item{q2SEd}{97.5% quantile (ie, +2SD) of population odds-ratios for subtype-differentiating SNPs. Corresponds to 'tau' in pars and overrides tau if set.}

\item{q2SEa1}{97.5% quantile (ie, +2SD) of population odds-ratios for disease-causative SNPs which do NOT differentiate subtypes (group 2). Corresponds to sigma_1 and overrides sigma_1 if set.}

\item{q2SEa2}{97.5% quantile (ie, +2SD) of population odds-ratios for disease-causative SNPs which DO differentiate subtypes (group 3). Corresponds to sigma_2 and overrides sigma_2 if set.}

\item{cor_st}{correlation, NOT covariance, between s2 and tau. Overrides pars if set.}

\item{seed}{random seed; if NULL is set to clock time}

\item{return_matrix}{if TRUE, returns a SNP matrix, otherwise returns Z_a and Z_d scores}

\item{null_model}{if pars=NULL and other parameters are null, parameters are chosen randomly. If null_model=TRUE, parameters are chosen from H0, otherwise H1.}
}
\value{
either object of type SnpMatrix, in which indices are in order controls, case subtype 1, case subtype 2; or Z_d and Z_a scores in an n x 2 matrix (Z[,1]=Z_d, Z[,2]=Z_a). Global variable pars_true contains 'true' values of parameters of Z_a, Z_d distribution and can be used to start the fitting algorithm. Global variable or_true contains values (pi0,pi1,q2SEd,q2SEa1,q2SEa2,cor_st)
}
\description{
Simulate a matrix of genotypes following a set of model parameters. Sets global variable pars_true containing 'true' parameter values, and or_true containing details of underlying odds ratio distribution.
}

