% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit.mml5.R
\name{summary.3Gfit}
\alias{summary.3Gfit}
\title{Summary method for class 3Gfit}
\usage{
summary.3Gfit(yy, n = 3, ...)
}
\arguments{
\item{yy}{object of class 3Gfit; generally output from fit.3g}

\item{n}{print this many rows of matrix}
}
\description{
Summary method for class 3Gfit
}
\author{
James Liley
}

