% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{summary.sim_N_test}
\alias{summary.sim_N_test}
\title{summary.Nsim_stat}
\usage{
summary.sim_N_test(yy)
}
\arguments{
\item{yy}{object of class sim_N_test, output from Nsim_stat}
}
\value{
for each bootstrap sample size, prints quantiles of each parameter
}
\description{
Summary method for class sim_N_test
}
\author{
James Liley
}

