% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulations.R
\name{summary.sim_output}
\alias{summary.sim_output}
\title{Summary method for class sim_output}
\usage{
summary.sim_output(yy)
}
\arguments{
\item{yy}{object of class sim_output; generally output from sim_analysis}
}
\description{
Summary method for class sim_output
}
\author{
James Liley
}

