% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.3g.R
\name{syscov}
\alias{syscov}
\title{syscov}
\usage{
syscov(v, n1, n2, nc)
}
\arguments{
\item{v}{the proportion of subgroup 1 in the population}

\item{n1}{number of samples of subgroup 1}

\item{n2}{number of samples of subgroup 2}

\item{nc}{number of controls}
}
\value{
estimated covariance
}
\description{
Determine correlation induced by adjusting Z scores for discrepancy in subgroup frequency
}
\details{
If a particular sub-trait within a disease has a frequency of \eqn{v} in the population, but is comparatively oversampled (or undersampled) in the case group to a frequency \eqn{n}, the observed effect size between cases and controls will be biased toward the effect size in the oversampled subgroup, compared to an observed effect size in a study where the subgroup frequency matches that of the population.

This can be corrected for by computing \eqn{Z_a' = Z_a + (v-n)Z_d}. However, this induces a systematic non-zero covariance between \eqn{Z_a} and \eqn{Z_d}.

This function evaluates the resultant covariance
}
\examples{
n1=100; n2=500; nc=1000; v=0.8
M=runif(1000,0,0.5) # population MAFs at 1000 SNPs
G1=rbinom(1000,2*n1,M)/(2*n1) # observed MAF at n1 (diploid) samples in subgroup 1
G2=rbinom(1000,2*n2,M)/(2*n2) # observed MAF at n2 (diploid) samples in subgroup 2
Gc=rbinom(1000,2*nc,M)/(2*nc) # observed MAF at nc (diploid) controls
Xd= G1 - G2 # adjusted MAF differences between case and control group 
Xa= (v*G1 + (1-v)*G2) - Gc # adjusted MAF differences between case and control group 
Zd= Xd/sqrt(M*(1-M)*(1/(2*n1) + 1/(2*n2))) # Z score, subgroup 1 vs subgroup 2
Za= Xa/sqrt(M*(1-M)*((v^2)/(2*n1) + ((1-v)^2)/(2*n2) + 1/(2*nc)))

cov(Zd,Za)
syscov(v,n1,n2,nc)
}
\author{
Chris Wallace and James Liley
}

