% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.3g.R
\name{wcor}
\alias{wcor}
\title{wcor}
\usage{
wcor(x, y, weights = rep(1, length(x)))
}
\arguments{
\item{x}{first vector}

\item{y}{second vector of same length as \code{\link{x}}}

\item{weights}{vector of weights of same length as \code{\link{x}} and \code{\link{y}}}
}
\description{
Finds a weighted correlation of two vectors \code{\link{x}} and \code{\link{y}}
}
\details{
Weighting procedure is equivalent to counting value \code{\{\link{x}[i],\link{y}[i]\}} with multiplicity \code{\link{weights}[i]}
}
\examples{
nn=100000
x=(1:10)/10 +runif(10); y=(1:10)/10 +runif(10)
mult=c(1:3,4:6,4:6,7:10,7:10,7:10); weights=c(1,1,1,2,2,2,3,3,3,3)
print(c(cor(x[mult],y[mult]),wcor(x,y,weights=weights)))
}
\author{
James Liley
}

