% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/processing.R
\name{z_scores}
\alias{z_scores}
\title{z_scores}
\usage{
z_scores(X, Ya, Yd, Ca = NULL, Cd = NULL, signed = TRUE, control = TRUE)
}
\arguments{
\item{X}{a snpMatrix object, QC'd, or a list of objects of type snpMatrix covering different SNPs (ie, one per chromosome)}

\item{Ya}{indices of cases/controls; Ya[i] corresponds to X[i,]. 0 for controls, 1 for cases}

\item{Yd}{indices of subgroups; Yd[i] corresponds to X[i,]. NA for controls, 1 for subtype 1, 2 for subtype 2, or the value of a quantitative trait across cases.}

\item{Ca}{matrix of covariates for computing Z_a; Ca[i,] gives the covariates for sample i corresponding to X[i,]}

\item{Cd}{matrix of covariates for computing Z_d; Cd[i,] gives the covariates for sample i corresponding to X[i,]. CD[i,] is ignored unless Ya[i]>0.}

\item{signed}{set to TRUE to return signed Z scores}

\item{control}{set to TRUE to perform genomic control on Z scores. We recommend strict genomic control (ie, so median(p)=0.5).}
}
\description{
Generates Z_a and Z_d scores from a QC'd genotype matrix
}
\examples{
# pending - need to find a SnpMatrix which can be made publically available
}
\author{
James Liley
}

