% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{za_scores}
\alias{za_scores}
\title{za_scores}
\usage{
za_scores(X, Ya, Ca = NULL, signed = TRUE, control = TRUE)
}
\arguments{
\item{X}{a snpMatrix object, QC'd, or a list of objects of type snpMatrix covering different SNPs (ie, one per chromosome)}

\item{Ya}{indices of cases/controls; Ya[i] corresponds to X[i,]. 0 for controls, 1 for cases}

\item{Ca}{matrix of covariates for computing Z_a; Ca[i,] gives the covariates for sample i corresponding to X[i,]}

\item{signed}{set to TRUE to return signed Z scores}

\item{control}{set to TRUE to perform genomic control on Z scores}
}
\description{
Generates Z_a scores from a QC'd genotype matrix
}
\examples{
# See examples for function z_scores
}
\author{
James Liley
}

