% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{zd_scores}
\alias{zd_scores}
\title{zd_scores}
\usage{
zd_scores(X, Ya, Ca = NULL, signed = TRUE, control = TRUE)
}
\arguments{
\item{X}{a snpMatrix object, QC'd, or a list of objects of type snpMatrix covering different SNPs (ie, one per chromosome)}

\item{signed}{set to TRUE to return signed Z scores}

\item{control}{set to TRUE to perform genomic control on Z scores##' @export}

\item{Yd}{indices of subgroups; Yd[i] corresponds to X[i,]. 1 for subtype 1, 2 for subtype 2, or the value of a quantitative trait across cases. MUST be NA for controls and cases for which subtype is unknown or undefined.}

\item{Cd}{matrix of covariates for computing Z_d; Cd[i,] gives the covariates for sample i corresponding to X[i,]. Cd[i,] is ignored if Yd[i]=NA}
}
\description{
Generates Z_d scores from a QC'd genotype matrix
}
\examples{
# See examples for function z_scores
}
\author{
James Liley
}
