% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit.mml5.R
\name{fit.3g}
\alias{fit.3g}
\title{fit.3g}
\usage{
fit.3g(Z, pars = c(0.8, 0.1, 2, 2, 3, 0.5), weights = rep(1, dim(Z)[1]),
  C = 1, fit_null = FALSE, maxit = 10000, tol = 1e-04, sg1 = FALSE,
  accel = TRUE, verbose = TRUE, file = NULL, n_save = 20,
  incl_z = TRUE)
}
\arguments{
\item{Z}{an n x 2 array; Z[i,1], Z[i,2] are the Z_d and Z_a scores respectively for the ith SNP}

\item{pars}{vector containing initial values of pi0,pi1,tau,sigma1,sigma2,rho.}

\item{weights}{SNP weights to adjust for LD; output from LDAK procedure}

\item{C}{a term C*log(pi0*pi1*pi2) is added to the likelihood so the model is specified.}

\item{fit_null}{set to TRUE to fit null model rho=0, tau=0}

\item{maxit}{maximum number of iterations before algorithm halts}

\item{tol}{how small a change in pseudo-likelihood halts the algorithm}

\item{sg1}{set to TRUE to force sigma1>=1, sigma2>=1, tau>=1.}

\item{accel}{attempts to accelerate the fitting process by taking larger steps.}

\item{verbose}{prints current parameters with frequency defined by 'n_save'}

\item{incl_z}{set to 'true' to include Z and weights in output. If 'false' these are set to null.}

\item{save}{history to a file with frequency defined by 'n_save'}

\item{b_int}{save or print pars every b_int iterations}
}
\value{
a list of six objects (class 3Gfit): pars is vector of fitted parameters, history is a matrix of fitted parameters and pseudo-likelihood at each stage in the E-M algorithm, logl is the joint pseudo-likelihood of Z_a and Z_d, logl_a is the pseudo-likelihood of Z_a alone (used for adjusting PLR),  z_ad is n x 2 matrix of Z_d and Z_a scores, weights is the weights used to generate the model, and hypothesis is 0 or 1 depending on the value of fit_null.
}
\description{
Fit a specific Gaussian mixture distribution. See main text for details.
}
\examples{
nn=100000
Z=abs(rbind(rmnorm(0.8*nn,varcov=diag(2)), rmnorm(0.15*nn,varcov=rbind(c(1,0),c(0,2^2))), rmnorm(0.05*nn,varcov=rbind(c(3^2,2),c(2,4^2))))); weights=runif(nn)
yy=fit.3g(Z,pars=c(0.7,0.2,2.5,1.5,3,1),weights=weights,incl_z=TRUE)
yy$pars
plot(yy,rlim=2)
}
\author{
Chris Wallace and James Liley
}

