% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_sim.R
\name{p2s}
\alias{p2s}
\title{p2s}
\usage{
p2s(p, n1, n2, mafs = NULL, NS = 50000, fast = TRUE)
}
\arguments{
\item{p}{standard deviation of underlying log odds ratio distribution}

\item{n1}{number of samples in group 1}

\item{n2}{number of samples in group 2}

\item{mafs}{use to specify distribution of average MAF (across both groups). Default is MAF~U(0.01,0.5)}

\item{NS}{number of SNPs to use in simulation}

\item{fast}{use quadratic approximation (TRUE) or simulate genotypes (FALSE)}
}
\value{
expected value of observed standard deviation of Z scores.
}
\description{
Converts underlying 'population' odds-ratio distribution into corresponding expected value of SD(Z) (ie, tau, sigma_1, or sigma_2). Assume that, across some set of SNPs in, population log-odds ratios between two phenotypes A and B are normally distributed with standard deviation p. If a GWAS is performed between a group of samples with phenotype A of size n1 and a group of samples of phenotype B of size n2, and Z-scores calculated for each SNP, this function returns the corresponding standard deviation for these Z scores.
}
\details{
Can either use a quick quadratic approximation (rapid, but dependent on MAF~U(0,0.5) and potentially inaccurate at small or very different n1,n2, or simulate genotypes (slow but accurate)
}
\examples{
p_underlying=0.15; # underlying standard deviation of odds-ratio distribution
nsnp=10000; # number of SNPs
n1=500; n2=1000 # number of samples in group 1, group 2

ors=exp(rnorm(nsnp,mean=0,sd=p_underlying)); # (simulated) underlying odds ratios

pmaf_1=runif(nsnp,0.05,0.5); # (simulated) underlying population minor allele frequencies in group 1
pmaf_2=ors*pmaf_1/(1-pmaf_1+(ors*pmaf_1)) # underlying population minor allele frequencies in group 2
plot(pmaf_2*(1-pmaf_1)/(pmaf_1*(1-pmaf_2)),ors); abline(0,1,col="red") # check

omaf_1=(rbinom(nsnp,n1,p=pmaf_1)+rbinom(nsnp,n1,p=pmaf_1))/(2*n1) # (simulated) observed diploid minor allele frequencies in group 1
omaf_2=(rbinom(nsnp,n2,p=pmaf_2)+rbinom(nsnp,n2,p=pmaf_2))/(2*n2) # (simulated) observed diploid minor allele frequencies in group 1

test=apply(cbind(omaf_1,omaf_2),1,function(x) prop.test(2*c(n1,n2)*x,2*c(n1,n2))$p.value) # chi-square tests between frequencies under H0: pmaf_1==pmaf_2

obs_z=-qnorm(test/2)*sign(omaf_2-omaf_1) # observed Z scores

sd(obs_z) # observed SD of Z scores
p2s(p_underlying,n1,n2) # predicted
}
