% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.3g.R
\name{plhood1}
\alias{plhood1}
\title{plhood1}
\usage{
plhood1(za, pars, weights = rep(1, length(za)), C = 1, sumlog = TRUE)
}
\arguments{
\item{za}{one dimensional vector of \eqn{Z_{a}}{Z_a} scores}

\item{pars}{vector of parameters (\code{pi0},\code{pi1},\code{tau},\code{sigma1},\code{sigma2},\code{rho})}

\item{weights}{LDAK weights corresponding to \eqn{Z_{a}}{Z_a}}

\item{C}{scaling factor for adjustment}

\item{sumlog}{set to TRUE to return the (weighted) sum of log pseudo-likelihoods for each datapoint; FALSE to return a vector of pseudo-likelihoods for each datapoint.}
}
\value{
value of pseudo- log likelihood of all observations (if \code{\link{sumlog}}==TRUE) or vector of n pseudo-likelihoods (if \code{\link{sumlog}}==FALSE)
}
\description{
Computes the likelihood of observed \eqn{Z_{a}}{Z_a}.
}
\details{
Used for establishing how much of a likelihood ratio is because \eqn{Z_{a}}{Z_a} is better-fitted by a three-Gaussian model than a two-Gaussian model. Strictly, calculates the expected value of \eqn{PL(Z_{d},Z_{a})}{PL(Z_d,Z_a)} when \eqn{Z_{d}}{Z_d} has uniform unit variance, and the parameter values \code{tau}, \code{rho} are set to 1, 0 respectively.
}
\author{
James Liley
}

