% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.3g.R
\name{plot.3Gfit}
\alias{plot.3Gfit}
\title{Plot method for class \code{\link{3Gfit}}.}
\usage{
\method{plot}{3Gfit}(yy, ...)
}
\arguments{
\item{yy}{object of class \code{\link{3Gfit}}; generally output from \code{\link{fit.3g}}}

\item{scales}{set to draw contours of the distributions at these levels (passed to \code{\link{plotpars}} and \code{\link{plotvdist}}).}
}
\description{
If the \code{z_ad} attribute of parameter \code{\link{yy}} is set, plots the absolute \eqn{Z_{a}}{Z_a} and \eqn{Z_{d}}{Z_d} scores (+/+ quadrant only), blocking out dense regions, and contours of the fitted Gaussians. Note that category 3 comprises two mirror-image Gaussians. Contours for category 1 are drawn in black, category 2 in blue, and category 3 in red.
}
\author{
James Liley
}
