% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.3g.R
\name{plotpars}
\alias{plotpars}
\title{plotpars}
\usage{
plotpars(pars, over = TRUE, ...)
}
\arguments{
\item{pars}{six element vector (\code{pi0},\code{pi1},\code{tau},\code{sigma1},\code{sigma2},\code{rho})}
}
\description{
Plots contour lines of three bivariate normal distributions as determined by \code{\link{pars}}
}
