% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit.mml5.R
\name{plotvdist}
\alias{plotvdist}
\title{lhood}
\usage{
plotvdist(varcov = diag(2), m = c(0, 0), scales = (1/(6.282 *
  det(varcov))) * (2^(-(1:5))), over = TRUE, xlim = NULL, ylim = NULL,
  ...)
}
\arguments{
\item{varcov}{covariance matrix (2D)}

\item{m}{mean vector}

\item{scales}{heights of normal pdf}

\item{over}{set to FALSE to draw new plot}

\item{xlim}{limits of x axis if over=FALSE}

\item{ylim}{limits of y axis if over=FALSE}
}
\description{
Plots contour lines of a bivariate normal distribution parametrised by covariance matrix, at specific distribution heights 'scales'
}

