% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.3g.R
\name{print.3Gfit}
\alias{print.3Gfit}
\title{Print method for class \code{\link{3Gfit}}}
\usage{
\method{print}{3Gfit}(y, n = 3, ...)
}
\arguments{
\item{yy}{object of class \code{\link{3Gfit}}; generally output from \code{\link{fit.3g}}}
}
\description{
Print method for class \code{\link{3Gfit}}
}
\author{
James Liley
}

