% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_sim.R
\name{sim_genotypes}
\alias{sim_genotypes}
\title{sim_genotypes}
\usage{
sim_genotypes(n_snps = 50000, n_control = 2000, n_case = c(1000, 1000),
  pars = NULL, q2SEd = NULL, q2SEa1 = NULL, q2SEa2 = NULL,
  cor_st = NULL, seed = NULL, return_matrix = FALSE, null_model = TRUE)
}
\arguments{
\item{n_snps}{number of SNPs}

\item{n_control}{number of controls}

\item{n_case}{two-element vector; number of cases in subtypes 1 and 2}

\item{pars}{expected observed parameter values. If NULL, values are chosen randomly; \code{pi0} from c(0.1,0.01,0.001),1), \code{pi1} from c(0.2,0.1,0.05); \code{q2SEd} from c(1,1.2,1.5,2); \code{q2SEa1} from c(1.2,1.5,2); (if \code{\link{null_model}}==FALSE) \code{q2SEa2} from c(1.2,1.5,2); and \code{rho} from c(0,0.1,0.5).}

\item{q2SEd}{97.5\% quantile (ie, +2SD) of population odds-ratios for subtype-differentiating SNPs. Corresponds to \code{tau} in \code{\link{pars}} and overrides \code{\link{pars}} if set.}

\item{q2SEa1}{97.5\% quantile (ie, +2SD) of population odds-ratios for disease-causative SNPs which do NOT differentiate subtypes (category 2). Corresponds to \code{sigma_1} in \code{\link{pars}} and overrides \code{\link{pars}} if set.}

\item{q2SEa2}{97.5\% quantile (ie, +2SD) of population odds-ratios for disease-causative SNPs which DO differentiate subtypes (category 3). Corresponds to \code{sigma_2} in \code{\link{pars}} and overrides \code{\link{pars}} if set.}

\item{cor_st}{correlation (as opposed to covariance) between \code{sigma2} and \code{tau}. Overrides \code{\link{pars}} if set.}

\item{seed}{random seed; if NULL is set to clock time}

\item{return_matrix}{if TRUE, returns a SNP matrix, otherwise returns Z_a and Z_d scores}

\item{null_model}{if \code{\link{pars}}=NULL and other parameters are null, parameters are chosen randomly. If \code{\link{null_model}}=TRUE, parameters are chosen from H0, otherwise H1.}
}
\value{
Either object of type SnpMatrix, in which indices are, in order: controls, case subtype 1, case subtype 2; or \eqn{Z_{d}}{Z_d} and \eqn{Z_{a}}{Z_a} scores in an n x 2 matrix, with Z[,1]=Z_d, Z[,2]=Z_a. Global variable \code{\link{pars_true}} contains 'true' values of parameters of \eqn{Z_{a}}{Z_a}, \eqn{Z_{d}}{Z_d} distribution and can be used to start the fitting algorithm. Global variable \code{\link{or_true}} contains values (\code{pi0},\code{pi1},\code{\link{q2SEd}},\code{\link{q2SEa1}},\code{\link{q2SEa2}},\code{\link{cor_st}})
}
\description{
Simulate a matrix of genotypes following a set of model parameters.
}
\details{
Sets global variable \code{\link{pars_true}} containing 'true' parameter values, and \code{\link{or_true}} containing details of underlying odds ratio distribution.
}
