% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.3g.R
\name{summary.3Gfit}
\alias{summary.3Gfit}
\title{Summary method for class \code{\link{3Gfit}}}
\usage{
\method{summary}{3Gfit}(yy, n = 3, ...)
}
\arguments{
\item{yy}{object of class \code{\link{3Gfit}}; generally output from \code{\link{fit.3g}}}

\item{n}{print this many rows of matrix}
}
\description{
Summary method for class \code{\link{3Gfit}}
}
\author{
James Liley
}
